requirements.txt
alpaca-py==0.38.0
protobuf==4.25.3
PyYAML==6.0.1
google-cloud-pubsub==2.28.0
google-cloud-logging==3.11.4